+++ packages/core/src/types/index.ts
@@ -1,3 +1,4 @@
 export * from './article'
+export * from './about'
 export * from './category'
 export * from './site-config'