+++ themes/react/src/App.tsx
@@ -6,6 +6,7 @@ import HomePage from './pages/HomePage'
 import LoginPage from './pages/LoginPage'
 import SearchPage from './pages/SearchPage'
 import ArchivesPage from './pages/ArchivesPage'
+import AboutPage from './pages/AboutPage'
 
 function AppLayout() {
   const location = useLocation()
@@ -25,6 +26,8 @@ function AppLayout() {
       document.title = `登录 - ${site}`
     } else if (p === '/archives') {
       document.title = `归档 - ${site}`
+    } else if (p === '/about') {
+      document.title = `关于我 - ${site}`
     } else if (p === '/search') {
       document.title = `搜索 - ${site}`
     } else {
@@ -55,6 +58,7 @@ function AppLayout() {
           <Route path="/login" element={<LoginPage />} />
           <Route path="/search" element={<SearchPage />} />
           <Route path="/archives" element={<ArchivesPage />} />
+          <Route path="/about" element={<AboutPage />} />
         </Routes>
       </div>