+++ themes/react/src/custom-elements.d.ts
@@ -9,6 +9,7 @@ declare global {
       'site-sidebar': CustomElement<HTMLElement> & { 'active-cat'?: string }
       'article-list': CustomElement<HTMLElement> & { 'active-cat'?: string }
       'archive-list': CustomElement<HTMLElement>
+      'about-page': CustomElement<HTMLElement>
       'article-viewer': CustomElement<HTMLElement>
       'search-panel': CustomElement<HTMLElement>
       'search-results': CustomElement<HTMLElement>