+++ themes/vue/src/router/index.ts
@@ -19,6 +19,11 @@ const router = createRouter({
       component: () => import('../views/ArchivesView.vue'),
     },
     {
+      path: '/about',
+      name: 'about',
+      component: () => import('../views/AboutView.vue'),
+    },
+    {
       path: '/search',
       name: 'search',
       component: () => import('../views/SearchView.vue'),